Stop using xml.etree
This is a bit extreme, but introduction of CDATA fully broke parsing
of existing XMLs. The right solution is to switch to lxml, but for now
fully hand-constructing the XML. The "advantage" is that namespaces
don't need any hacky handling. The most brittleness is in
_get_current_rss_items, which looks for '<item>' as the boundary of
items. This is okay for all the feeds of current interest. It can
be easily broken, but the positive is that any breaks will be limited
to that user's feed.

Also using longer sleeps between API calls to minimize rate-limit errors
which have already started showing up.