Support for daily digest
If a non-None Config.DAILY_DIGEST is specified, on each day a little after that time, all tweets
since the previous RSS update will be batched into a single item.

So this is not strictly a daily digest, and the first run when the app is started may batch
tweets sent in more than a single day.